changed type for price

+++ task.sql
@@ -6,7 +6,7 @@ CREATE TABLE Products (
     ID INT AUTO_INCREMENT,
     Name VARCHAR(50),
     Description VARCHAR(100),
-    Price FLOAT,
+    Price INT,
     WarehouseAmount INT,
     PRIMARY KEY (ID)
 );
